+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Guid PKs will have a default value of `Guid.NewGuid()` in their creation dto -- fixes #28
 - PK already exists guard will be performed when adding a new entity and throw a 409 conflict via a new conflict exception if a record already exists with that guid. -- fixes #29
   - **Add an integ test for this**
+- Updated `ProducesResponseType` in controllers to generic `Response` type where applicable
 
 ## [0.9.3] - 2021-04-10